+++ CHANGELOG.md
@@ -11,7 +11,9 @@ Notable changes:
   * `ubik`:
     * Finds functions from the library, not just commands (scripts).
     * Smarter error reporting.
-  * `progress-msg` defaults to enabled.
+  * `stderr`
+    * New helper function to define standardized control options.
+    * progress-msg` defaults to enabled.
 * `bashy-basics`:
   * New script `timey secs`. Subsumes and deprecates `timey now-secs`.
   * New helper functions `jbash-array` and `jpostproc`.